Fix problems with chats sessions lifecycles
