+++ src/main/kotlin/com/sourcegraph/cody/chat/AgentChatSessionService.kt
@@ -1,0 +1,45 @@
+package com.sourcegraph.cody.chat
+
+import com.intellij.openapi.components.Service
+import com.intellij.openapi.components.service
+import com.intellij.openapi.project.Project
+import com.sourcegraph.cody.agent.CodyAgent
+import com.sourcegraph.cody.history.state.ChatState
+import java.util.concurrent.ConcurrentLinkedQueue
+
+@Service(Service.Level.PROJECT)
+class AgentChatSessionService(private val project: Project) {
+  private val chatSessions: ConcurrentLinkedQueue<AgentChatSession> = ConcurrentLinkedQueue()
+
+  fun addSession(session: AgentChatSession) {
+    chatSessions.add(session)
+  }
+
+  fun removeAllSessions() {
+    chatSessions.clear()
+  }
+
+  fun removeSession(state: ChatState): Boolean {
+    val session = chatSessions.find { it.getInternalId() == state.internalId }
+    return chatSessions.remove(session)
+  }
+
+  fun getOrCreateFromState(state: ChatState): AgentChatSession {
+    val session = chatSessions.find { it.getInternalId() == state.internalId }
+    return session ?: AgentChatSession.createFromState(project, state)
+  }
+
+  fun getSession(sessionId: SessionId): AgentChatSession? =
+      chatSessions.find { it.hasSessionId(sessionId) }
+
+  fun restoreAllSessions(agent: CodyAgent) {
+    chatSessions.forEach { it.restoreAgentSession(agent) }
+  }
+
+  companion object {
+    @JvmStatic
+    fun getInstance(project: Project): AgentChatSessionService {
+      return project.service<AgentChatSessionService>()
+    }
+  }
+}
