+++ src/main/kotlin/com/sourcegraph/cody/chat/NewChatAction.kt
@@ -9,7 +9,7 @@ class NewChatAction : DumbAwareAction() {
   override fun actionPerformed(event: AnActionEvent) {
     val project = event.project ?: return
     CodyToolWindowContent.executeOnInstanceIfNotDisposed(project) {
-      addChatSession(AgentChatSession.createNew(project))
+      switchToChatSession(AgentChatSession.createNew(project))
     }
   }
 }
