+++ src/main/kotlin/com/sourcegraph/cody/commands/ui/CommandsContextMenu.kt
@@ -33,7 +33,8 @@ class CommandsContextMenu {
                   CodyEditorFactoryListener.Util.informAgentAboutEditorChange(
                       it, hasFileChanged = false) {
                         CodyToolWindowContent.executeOnInstanceIfNotDisposed(project) {
-                          addChatSession(AgentChatSession.createFromCommand(project, commandId))
+                          switchToChatSession(
+                              AgentChatSession.createFromCommand(project, commandId))
                         }
                       }
                 }
