+++ src/main/kotlin/com/sourcegraph/cody/history/state/MessageState.kt
@@ -9,7 +9,8 @@ class MessageState : BaseState() {
 
   @get:OptionTag(tag = "text", nameAttribute = "") var text: String? by string()
 
-  @get:OptionTag(tag = "speaker", nameAttribute = "") var speaker: SpeakerState? by enum<SpeakerState>()
+  @get:OptionTag(tag = "speaker", nameAttribute = "")
+  var speaker: SpeakerState? by enum<SpeakerState>()
 
   // todo var contextFiles by list<String>()
